+++ intelmq/bots/collectors/http/collector_http.py
@@ -11,14 +11,15 @@ http_verify_cert: boolean
 extract_files: value used to extract files from downloaded compressed file
     default: None
     all: True; some: string with file names separated by ,
+http_url_formatting: bool|json to turn on time formatting (and to specify delta to current time)
 http_username, http_password: string
 http_proxy, https_proxy: string
 http_timeout_sec: tuple of two floats or float
 http_timeout_max_tries: an integer depicting how often a connection attempt is retried
 """
-import datetime
 import io
 import zipfile
+from datetime import datetime, timedelta
 
 try:
     import requests
@@ -30,8 +31,14 @@ from intelmq.lib.utils import unzip
 
 
 class Time(object):
+    def __init__(self, delta=None):
+        """ Delta is a datetime.timedelta JSON string, ex: '{days=-1}'. """
+        self.time = datetime.now()
+        if not isinstance(delta, bool):
+            self.time += timedelta(**delta)
+
     def __getitem__(self, timeformat):
-        return datetime.datetime.now().strftime(timeformat)
+        return self.time.strftime(timeformat)
 
 
 class HTTPCollectorBot(CollectorBot):
@@ -44,8 +51,17 @@ class HTTPCollectorBot(CollectorBot):
         self.extract_files = getattr(self.parameters, "extract_files", None)
 
     def process(self):
-        if getattr(self.parameters, 'http_url_formatting', False):
-            http_url = self.parameters.http_url.format(time=Time())
+
+        formatting = getattr(self.parameters, 'http_url_formatting', False)
+        if formatting:
+            try:
+                http_url = self.parameters.http_url.format(time=Time(formatting))
+            except TypeError:
+                self.logger.error("Wrongly formatted http_url_formatting parameter: %s. Should be boolean or a time-delta JSON.", formatting)
+                raise
+            except KeyError:
+                self.logger.error("Wrongly formatted http_url parameter: %s. Possible misspell with 'time' variable.", self.parameters.http_url)
+                raise
         else:
             http_url = self.parameters.http_url